% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_path_mat.R
\name{gen_path_mat}
\alias{gen_path_mat}
\title{Generate path matrix from data file}
\usage{
gen_path_mat(x, diag = T)
}
\arguments{
\item{x}{The data file}

\item{diag}{Whether diagonal elements should exist}
}
\value{
Empty matrix with named rows and columns corresponding to individual data files
}
\description{
Helper function to generate an empty similarity matrix based on 
header of individual data file
}
\examples{
gen_path_mat(data.frame(V1 = rnorm(10, 0, 1), V2 = rnorm(10, 0, 1)))
}
