% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimmeGUI.R
\name{gimmeGUI}
\alias{gimmeGUI}
\title{Graphical User Interface for the gimme R Package}
\description{
A graphical user interface for the gimme R package. Easily facilitates the specification of directories, custom subgroups, fixed and freed paths for estimation, and various other options.
}
\examples{
\dontrun{
gimmeGUI()
}
}
\references{
Lane, S.T. & Gates, K.M. (2017). Automated selection of robust
individual-level structural equation models for time series data.
Structural Equation Modeling, 24, 768-782.

Gates, K.M. & Molenaar, P.C.M. (2012). Group search algorithm
recovers effective connectivity maps for individuals
in homogeneous and heterogeneous samples. NeuroImage, 63, 310-319.
}
